% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openwater.R
\name{ow_get_api}
\alias{ow_get_api}
\title{OpenWater Get Application Programming Interface}
\usage{
ow_get_api(
  ow_api_key = Sys.getenv("OPENWATER_API_KEY"),
  ow_domain = Sys.getenv("OPENWATER_DOMAIN")
)
}
\arguments{
\item{ow_api_key}{OpenWater API Key}

\item{ow_domain}{OpenWater domain}
}
\value{
An object representing the OpenWater API operations and schemas
}
\description{
OpenWater Get Application Programming Interface
}
\examples{
\donttest{ow_api <- ow_get_api()}
}
