% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openwater.R
\name{ow_get_report}
\alias{ow_get_report}
\title{OpenWater Get Report}
\usage{
ow_get_report(reportId, ow_api, sleep = 3, ...)
}
\arguments{
\item{reportId}{ID of the desired report}

\item{ow_api}{OpenWater API object}

\item{sleep}{The time interval to wait between checking job completion status}

\item{...}{Additional arguments passed to \code{\link[readr:read_delim]{read_csv}}()}
}
\value{
A \code{\link[tibble]{tibble}}
}
\description{
OpenWater Get Report
}
\examples{
\donttest{ow_get_report(666, ow_api)}
}
